requirements.txt
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysql-connector==2.2.9
mysql-connector-python==8.0.22
protobuf==3.14.0
simplejson==3.17.2
six==1.15.0
Werkzeug==1.0.1
